requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.22.0
urllib3==1.26.13
